+++ chakracore-cxx/lib/Runtime/Language/arm64/StackFrame.h
@@ -4,6 +4,8 @@
 //-------------------------------------------------------------------------------------------------------
 #pragma once
 
+#include "Language/StackFrameConstants.h"
+
 /*
  * Stackwalking on Evanesco:
  * ------------------------
