+++ chakracore-cxx/lib/Runtime/Language/i386/StackFrame.h
@@ -4,6 +4,8 @@
 //-------------------------------------------------------------------------------------------------------
 #pragma once
 
+#include <stdint.h>
+#include "Language/StackFrameConstants.h"
 
 namespace Js {
